feat: update API

+++ src/main/java/com/msig/phonebook/PhoneBookRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 
-public interface PhoneBookRepository extends JpaRepository<Phonebook, Integer> {
+public interface PhoneBookRepository extends JpaRepository<Phonebook, Long> {
     List<Phonebook> findByNameContaining(String name);
 }
